refactor(exchange): entities renamed to functions, extract all snapshot related entities to own dir, rename getters to getOrCreate
