+++ subgraphs/exchange/src/functions/index.ts
@@ -1,11 +1,8 @@
 export * from './bundle'
 export * from './factory'
 export * from './liquidity-position'
-export * from './liquidity-position-snapshot'
 export * from './pair'
-export * from './pair-day-data'
-export * from './pair-hour-data'
+export * from './snapshots'
 export * from './token'
-export * from './token-day-data'
 export * from './user'
