+++ subgraphs/exchange/src/functions/snapshots/pair-day-data.ts
@@ -1,7 +1,7 @@
 import { BigInt, ethereum } from '@graphprotocol/graph-ts'
-import { PairDayData } from '../../generated/schema'
-import { BIG_DECIMAL_ZERO } from '../constants'
-import { getOrCreatePair } from './pair'
+import { PairDayData } from '../../schema'
+import { BIG_DECIMAL_ZERO } from '../../constants'
+import { getOrCreatePair } from '../pair'
 
 export function updatePairDayData(event: ethereum.Event): PairDayData {
   const timestamp = event.block.timestamp.toI32()
