+++ subgraphs/exchange/src/mappings/factory.ts
@@ -1,10 +1,10 @@
 import { PairCreated } from '../../generated/Factory/Factory'
 import { Pair as PairTemplate } from '../../generated/templates'
 import { BIG_INT_ONE } from '../constants'
-import { getFactory, getOrCreatePair } from '../enitites'
+import { getOrCreateFactory, getOrCreatePair } from '../functions'
 
 export function onPairCreated(event: PairCreated): void {
-  const factory = getFactory()
+  const factory = getOrCreateFactory()
 
   const pair = getOrCreatePair(event.params.pair, event.block)
